+++ src/main/java/org/kdb/inside/brains/view/export/FlipTableExportAction.java
@@ -12,8 +12,8 @@ import org.jetbrains.annotations.Nullable;
 import org.kdb.inside.brains.core.KdbQuery;
 import org.kdb.inside.brains.core.KdbResult;
 import org.kdb.inside.brains.view.KdbOutputFormatter;
-import org.kdb.inside.brains.view.console.TableResult;
-import org.kdb.inside.brains.view.console.TableResultView;
+import org.kdb.inside.brains.view.console.table.TableResult;
+import org.kdb.inside.brains.view.console.table.TableResultView;
 
 import javax.swing.*;
 import java.awt.*;